// src/features/products/types/product.d.ts
interface Product{
    sku:number ,
    name:string ,
    category:name ,
    image:string , 
    basePrice:number ,
    lastUpdate:string ,



}



export interface ProductRequest{
    sku:number ,
    name:string ,
    category:name ,
    image:string , 
    basePrice:number ,
    requestDate:string ,
    vendor:string

}